Add support for nested hierarcies.
This is only lightly tested-- our kernel doesn't yet properly export
statistics for nested kernels, but the groups seem to end up generally
in the right task.

One problem is that the task for "t0/t1" in io ends up in "t1" for the
cpu, not "t0". This isn't a problem for the simplest test, but we should
fix it soon.