Score experiments in terms of total weight.
Instead of scoring out of a max of 1000, time measurements are now in
terms of the total assigned weight. This allows us to do proportional
testing without requiring all the containers to add up to 1000.

This should give us better coverage for "slack" testing.

Additionally, add output to an autotest file with the '-o' flag.

This changes our test output!

Here's a test run with this config:
Running experiment 0: 500 rdrand, 100 (100 rdrand, 100 rdrand, 100 rdrand)
...
experiment 0 achieved DTFs: 501, 98 [100, 102, 97]

Note that the DTFs match the experiment string, but they're not out of a
total of 1000 anymore.