Added interface to define container priorties.
The tests currently only accept a weight for a group, this patch
allows for a priority to be specified along with the weight. This
is done by appending the character 'P' to the weight to specify
high priority and just a weight in the regular case.

For example '600P rdrand, 400 wrseq.dir' would mean that the
first group would have a weight of 600 and also be high priority.

Change-Id: Ib0782c96abe71f7447d598818c66d0ba04fdc386