Rescore failed experiments using io.timeslice_used
There are workloads that cause unaccounted_sl to
build up quite a bit and so the service time and the
actual charged timeslices diverge quite a lot. This
has the capability to mask actual problems with scheduling.

This patch makes the test first score the max error
using service time. If that fails, it emits a warning that
the experiment is being re-scored and then re-scores it
using io.timeslice_used. The rescored numbers are for human
consumption only, it does not change the pass/fail status
or write anything to the autotest/csv output file.

Change-Id: I1dc614d769d00b5aaf64edf2d5860d3da39b7c8b