Move autotest file generation to the end.
Child processes were flushing lines intended for the autotest output
file, so duplicate entries were reported in test runs. The simplest way
to address this is to wait until we aren't going to run any more
experiments, and just output the data then.

I tested this with simple_nested_test and verified that the output is
sane.