Add a priority test, print long wait percentages.
This introduces "prio_test" to do some simple early prio testing, and
summarizes some of the data from I/O wait times.

This also brings back the flexible "post_experiment_cb" callback to
print container info at the end of an experiment.

Example output:
When a container has a worker string, we output it for readability:
14:21:24 INFO container blkcgroupt0 (io_load_read) Read wait: 28.000000% > 50ms, 28.000000% > 100ms
14:21:24 INFO container blkcgroupt1 (rdrand) Read wait: 41.935484% > 50ms, 23.655914% > 100ms