Calculate & log total throughput for experiments.
Calculate and log the aggregate throughput for all the experiments
run. The throughput is calculated by looking at the value of
io.io_service_bytes for the device at the root cgroup and dividing
this by the number of seconds that the experiment was run. For now
we simply log the value calculated at the end of the experiment.
Note that the value of the throughput has no bearing on whether the
test passes or fails.

The end result is that the log file looks similar to this:
...
15:40:06 INFO All workers have now completed or been killed by fastest worker.
15:40:06 INFO Experiment completed in 23.4 seconds
15:40:06 INFO Aggregate Throughput = 71.453175 MB/s
15:40:06 INFO experiment 4 achieved DTFs: 586, 204, 208
15:40:06 INFO experiment 4 PASSED: max observed error is 14, allowed is 35